requirement.txt
Flask==0.10.1
Flask-Bootstrap==3.3.0.1
Flask-HTTPAuth==2.3.0
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Moment==0.4.0
Flask-SQLAlchemy==2.0
Flask-SSLify==0.1.4
Flask-Script==2.0.5
Flask-WTF==0.10.3
ForgeryPy==0.1
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
MySQL-python==1.2.5
Pygments==2.0.2
SQLAlchemy==0.9.8
WTForms==2.0.2
Werkzeug==0.9.6
alembic==0.7.4
argparse==1.2.1
bleach==1.4.1
blinker==1.3
colorama==0.3.3
gevent==1.0.1
greenlet==0.4.5
gunicorn==18.0
html5lib==0.999
httpie==0.8.0
ipython==2.3.1
itsdangerous==0.24
meld3==1.0.0
python-openid==2.2.5
requests==2.5.1
selenium==2.44.0
six==1.9.0
supervisor==3.1.3
wsgiref==0.1.2
